requirements-dev.txt
Flask==2.0
Pillow==9.2.0
mock==4.0.3
black==22.3.0
isort==5.10.1
pylint==2.14.4
yamllint==1.26.3
